added std_lib_facilities.h
std_lib_facilities.h is described in the book as:
"std_lib_facilities.h. We wrote that file to simplify use of the
facilities available in all implementations of C++
(“the C++ standard library”). We will explain its contents as
we go along. It is perfectly ordinary standard C++,
but it contains details that we’d rather not bother you with
for another dozen chapters."